Add Jest, jest-cli, and rename components and add test files
