+++ client/src/components/DailyWeather/index.jsx
