+++ client/src/components/LocationPin/index.jsx
